# Make to_json twice as fast as it used to be