@bf4 2018-04-22
@shishirmk I think there's a lot to discuss in this PR.

Firstly, is the argument that the *first* time the method is called is twice as fast?  If so, I don't think that's a useful benchmark. Mostly applications will live to have the method called more than that.

For what it's worth, this pattern comes from a perf PR to ActiveModelSerializers that still holds up https://github.com/rails-api/active_model_serializers/pull/233

The rest of the changes in this PR need a discussion of cost benefit analysis, including taking into account that the original PR was merged without a test suite, but was tested otherwise.  The various methods removed in this PR are critical to configuring and debugging the behavior.  I would use it in https://github.com/Netflix/fast_jsonapi/issues/135