@bf4 2018-04-22
On my machine, a MBP running Ruby 2.4.1 I run the bench against master and get about 0.28ms each run and against make_to_json_fast about 0.11 ms each run.

But, if I change the benchmark to run to_json just once, I get the same 0.03 ms on each branch.  

```diff
require 'spec_helper'

describe FastJsonapi::MultiToJson do
  include_context 'movie class'

  before(:all) { GC.disable }
  after(:all) { GC.enable }

  describe 'self.to_json' do
    it do
+ FastJsonapi::MultiToJson.to_json movie
      second = Benchmark.measure { FastJsonapi::MultiToJson.to_json movie }.real * 1000
      print 'to_json: '
      puts "#{second} ms"
    end
  end
end
```

Let's rewrite the benchmark to use `Benchmark.bmbm`

> Sometimes benchmark results are skewed because code executed earlier encounters different garbage collection overheads than that run later. bmbm attempts to minimize this effect by running the tests twice, the first time as a rehearsal in order to get the runtime environment stable, the second time for real. GC.start is executed before the start of each of the real timings; the cost of this is not included in the timings. In reality, though, there’s only so much that bmbm can do, and the results are not guaranteed to be isolated from garbage collection and other effects.

```ruby
require 'spec_helper'

describe FastJsonapi::MultiToJson do
  include_context 'movie class'

  describe 'self.to_json' do
    it do
      n = 1_000
      Benchmark.bmbm do |x|
        x.report('to_json') {
          i = 0
          while i < n
            FastJsonapi::MultiToJson.to_json movie
            i += 1
          end
        }
      end
    end
  end
end
```

Here are the outputs against each branch.  I'm not going to label which branch is which 'cause they're the same :)

```
Rehearsal -------------------------------------------
to_json   0.020000   0.000000   0.020000 (  0.032862)
---------------------------------- total: 0.020000sec

              user     system      total        real
to_json   0.020000   0.000000   0.020000 (  0.021913)
```.

```
📝 $ be rspec spec/bench_spec.rb
Rehearsal -------------------------------------------
to_json   0.030000   0.010000   0.040000 (  0.029570)
---------------------------------- total: 0.040000sec

              user     system      total        real
to_json   0.020000   0.000000   0.020000 (  0.022821)
```