@shuheiktgw 2018-04-22
@bf4 Thank you so much for your comments! You are absolutely right if you take a benchmark multiple time the result will be the same since the difference appears only for the first time that `to_json` is called. 

Even if this PR does not improve performance, I personally believe that I can simply `to_json` method by using `class_eval`. What would you think if I make back some logging features I removed?